Uploaded Game
Uploaded the Source code of the Hamlet Game. This game was made in about
100 hours for an English creative assignment. I was working on something
different, but changed this which is why some things are called
BloodNight instead of Hamlet. I was rushed so lots of the code is
uncommented and messy and parts of it are weird such as event handling
through TileMap and TextBox